extract ol code
